test: organize tests to make room for unit tests
