+++ jest.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   testMatch: [
-    '<rootDir>/test/*.test.js'
+    '<rootDir>/test/unit/**/*.test.js',
+    '<rootDir>/test/acceptance/**/*.test.js',
   ],
   modulePathIgnorePatterns: [
     '<rootDir>/test/.*fixtures/*'
